pyproject.toml
[project]
name = "aerosentry-ai"
version = "1.0.0"
description = "AI-powered aviation security system using Software Defined Radio"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AeroSentry AI Team"}
]
keywords = ["aviation", "security", "sdr", "adsb", "anomaly-detection", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "asyncpg>=0.28.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "protobuf>=4.0.0",
]

[project.optional-dependencies]
edge = [
    "pyrtlsdr>=0.3.0",
    "pyarrow>=12.0.0",
]
cloud = [
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "minio>=7.0.0",
    "nats-py>=2.0.0",
]
ml = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "openai-whisper>=20230314",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.scripts]
aerosentry-edge = "edge.main:main"
aerosentry-cloud = "cloud.main:main"
aerosentry-setup-db = "scripts.setup_db:main"
aerosentry-evaluate = "scripts.run_evaluation:main"

[project.urls]
Homepage = "https://github.com/aerosentry/aerosentry-ai"
Documentation = "https://github.com/aerosentry/aerosentry-ai#readme"
Issues = "https://github.com/aerosentry/aerosentry-ai/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["edge*", "cloud*", "shared*", "disaster*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]

[tool.isort]
profile = "black"
line_length = 100
